split everything in correct files
