+++ examples/arctic_basin_seasonal_cycle.jl
@@ -184,7 +184,7 @@ Q = [datum[5] for datum in series]
 
 column_grid = RectilinearGrid(size=16, z=SeaIceColumnDiscretization((0, 4)), topology=(Flat, Flat, Bounded))
 
-column_thermodynamics = prescribed_salinity_enthalpy_thermodynamics(column_grid;
+column_thermodynamics = evolving_salinity_mushy_thermodynamics(column_grid;
     salinity_profile = 0,
     energy_transport = ConductiveTemperatureTransport(conductivity = 2),
     heat_boundary_conditions = (top = MeltingConstrainedFluxBalance(),
